% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{sym_fred}
\alias{sym_fred}
\alias{import_series.fred}
\title{Symbol specification for FRED}
\usage{
sym_fred(symbols, ..., curl_options = list())

\method{import_series}{fred}(symbol_spec, dates = NULL, ...)
}
\arguments{
\item{symbols}{Series symbols to import.}

\item{\dots}{Arguments passed to other functions (not currently used).}

\item{curl_options}{Options passed to \pkg{curl} functions.}

\item{symbol_spec}{A \code{symbol_spec} object, with one element for each
symbol that will be imported.}

\item{dates}{An ISO-8601 string specifying the start and/or end dates.}
}
\value{
An object of class \code{symbol_spec}. The object contains a
    \code{src_attr} attribute that stores any additional details that need
    to be used for the connection to the data source.

A \code{xts} object, with all the columns for each
   \code{symbol_spec}.
}
\description{
Imports data from FRED for each symbol in \code{symbol_spec}. This method
should not be called directly. Use \code{import_series} with one FRED
\code{symbol_spec}.
}
\examples{

series_symbols <- sym_fred(c("DGS10", "DGS5"))

\dontrun{
# one symbol
treasury_10y <- import_series(sym_fred("DGS10"))

# multiple symbols (one wide xts object)
treasury_rates <- import_series(series_symbols)
}

}
\references{
\url{https://fred.stlouisfed.org/}
}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
