% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_collection.R
\name{import_collection}
\alias{import_collection}
\title{Import a collection of data from one or more sources}
\usage{
import_collection(symbol_spec, dates = NULL, ...)
}
\arguments{
\item{symbol_spec}{A symbol specification object.}

\item{dates}{An ISO-8601 string specifying the start and/or end dates.}

\item{\dots}{Additional parameters passed to methods.}
}
\value{
An object of class \code{list}.
}
\description{
This function imports multiple data series from one or more sources. Each
series may have different column names.
}
\examples{

\dontrun{
series_symbols <- sym_fred(c("DGS10", "DGS5"))
collection <- import_collection(series_symbols)
}

}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
