% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ohlc_tiingo.R
\name{import_multi_ohlc.tiingo}
\alias{import_multi_ohlc.tiingo}
\alias{import_ohlc.tiingo}
\alias{import_ohlc_tiingo}
\title{Import data from Tiingo}
\usage{
\method{import_multi_ohlc}{tiingo}(symbol_spec, ..., from = NULL, to = NULL)
}
\arguments{
\item{symbol_spec}{A \code{symbol_spec} object, with one element for each
symbol that will be imported.}

\item{\dots}{Arguments passed to other functions (not currently used).}

\item{from}{A date/time that specifies the first possible date of the
imported data (default: \code{Sys.Date() - 365}).}

\item{to}{A date/time that specifies the last possible date of the
imported data (default: \code{Sys.Date()}).}
}
\value{
A \code{multiple_ohlc} object, with one element for each
   \code{symbol_spec}.
}
\description{
Imports data from Tiingo for each symbol in \code{symbol_spec}. This method
should not be called directly. Use \code{import_ohlc} with one or more
Tiingo \code{symbol_spec}.
}
\examples{

 ### Note: you must have a working internet
 ### connection for these examples to work!
 if (interactive()) {
     sym_spec <- sym_tiingo(c("IBM", "CSCO"), api_key = "[your_api_key]")
     tiingo_data <- import_ohlc(sym_spec)
 }

}
\keyword{IO}
\keyword{connection}
