% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ohlc.R
\name{import_ohlc}
\alias{import_ohlc}
\title{Import a single OHLC(VA) series as an xts object}
\usage{
import_ohlc(symbol_spec, dates = NULL, ...)
}
\arguments{
\item{symbol_spec}{A symbol specification object with one symbol.}

\item{dates}{An ISO-8601 string specifying the start and/or end dates.}

\item{\dots}{Additional parameters passed to methods.}
}
\value{
An object of class \code{xts}.
}
\description{
This function imports data that has columns with open, high, low, close,
and possibly volume and/or adjusted close (if the data source returns it).
}
\examples{

ticker <- sym_yahoo("AAPL")
ohlc <- import_ohlc(ticker)

}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
