% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_frame.R
\name{make_xts_frame}
\alias{make_xts_frame}
\title{Make an xts object from a multiple_ohlc object}
\usage{
make_xts_frame(multi_ohlc, column = "close", ...)
}
\arguments{
\item{multi_ohlc}{A \code{multiple_ohlc} object.}

\item{column}{Column to extract for each symbol.}

\item{\dots}{Not currently used.}
}
\value{
An object of class \code{xts}.
}
\description{
This function extracts a single column from each symbol in a multiple_ohlc
object.
}
\examples{

\dontrun{
    tickers <- sym_yahoo(c("AAPL", "NFLX"))
    ohlc <- import_ohlc(tickers)
    prices <- make_xts_frame(ohlc, "close")
}

}
\author{
Joshua Ulrich
}
