% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_spec.R
\name{sym_fred}
\alias{sym_fred}
\title{Symbol specification for FRED}
\usage{
sym_fred(symbols, ..., curl_options = list())
}
\arguments{
\item{symbols}{Series symbols to import.}

\item{\dots}{Additional source attributes.}

\item{curl_options}{Options passed to \pkg{curl} functions.}
}
\value{
An object of class \code{symbol_spec}. The object contains a
    \code{src_attr} attribute that stores any additional details that need
    to be used for the connection to the data source.
}
\description{
Symbol specification for FRED
}
\examples{

series_symbols <- sym_fred(c("DGS10", "DGS5"))

\dontrun{
    treasury_rates <- import_series(series_symbols)
}

}
\references{
\url{https://fred.stlouisfed.org/}
}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
