% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_spec.R
\name{sym_tiingo}
\alias{sym_tiingo}
\title{Symbol specification for Tiingo}
\usage{
sym_tiingo(symbols, ..., curl_options = list(), api_key = NULL)
}
\arguments{
\item{symbols}{Ticker symbols to import.}

\item{\dots}{Additional source attributes.}

\item{curl_options}{Options passed to \pkg{curl} functions.}

\item{api_key}{Your Tiingo API key (available for free).}
}
\value{
An object of class \code{symbol_spec}. The object contains a
    \code{src_attr} attribute that stores any additional details that need
    to be used for the connection to the data source.
}
\description{
Symbol specification for Tiingo
}
\examples{

\dontrun{
    tickers <- sym_tiingo(c("AAPL", "NFLX"), api_key = "*****")
    ohlc <- import_ohlc(tickers)
}

}
\references{
\url{https://www.tiingo.com/}
}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
