% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_spec.R
\name{sym_yahoo}
\alias{sym_yahoo}
\title{Symbol specification for Yahoo Finance}
\usage{
sym_yahoo(symbols, ..., curl_options = list())
}
\arguments{
\item{symbols}{Ticker symbols to import.}

\item{\dots}{Additional source attributes.}

\item{curl_options}{Options passed to \pkg{curl} functions.}
}
\value{
An object of class \code{symbol_spec}. The object contains a
    \code{src_attr} attribute that stores any additional details that need
    to be used for the connection to the data source.
}
\description{
Symbol specification for Yahoo Finance
}
\examples{

tickers <- sym_yahoo(c("AAPL", "NFLX"))

}
\references{
\url{https://finance.yahoo.com/}
}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
