% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiingo.R
\name{sym_tiingo}
\alias{sym_tiingo}
\alias{import_ohlc_collection.tiingo}
\alias{import_ohlc.tiingo}
\title{Symbol specification for Tiingo OHLC data}
\usage{
sym_tiingo(symbols, ..., curl_options = list(), api_key = NULL)

\method{import_ohlc_collection}{tiingo}(symbol_spec, dates = NULL, ...)
}
\arguments{
\item{symbols}{Ticker symbols to import.}

\item{\dots}{Arguments passed to other functions (not currently used).}

\item{curl_options}{Options passed to \pkg{curl} functions.}

\item{api_key}{Your Tiingo API key (available for free).}

\item{symbol_spec}{A \code{symbol_spec} object, with one element for each
symbol that will be imported.}

\item{dates}{An ISO-8601 string specifying the start and/or end dates.}
}
\value{
An object of class \code{symbol_spec}. The object contains a
    \code{src_attr} attribute that stores any additional details that need
    to be used for the connection to the data source.

A \code{ohlc_collection} object, with one element for each
   \code{symbol_spec}.
}
\description{
Imports data from Tiingo for each symbol in \code{symbol_spec}. This method
should not be called directly. Use \code{import_ohlc_collection} with one or
more Tiingo \code{symbol_spec}.
}
\examples{

tickers <- sym_tiingo(c("AAPL", "NFLX"), api_key = "*****")

\dontrun{
# one symbol
spy <- import_ohlc(sym_tiingo("SPY"))

# multiple symbols
ohlc <- import_ohlc_collection(tickers)
}


 ### Note: you must have a working internet
 ### connection for these examples to work!
 if (interactive()) {
     sym_spec <- sym_tiingo(c("IBM", "CSCO"), api_key = "[your_api_key]")
     tiingo_data <- import_ohlc_collection(sym_spec)
 }

}
\references{
\url{https://www.tiingo.com/}, \url{https://api.tiingo.com/}
}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
