% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo.R
\name{sym_yahoo}
\alias{sym_yahoo}
\alias{import_ohlc_collection.yahoo}
\alias{import_ohlc.yahoo}
\title{Symbol specification for Yahoo Finance OHLC data}
\usage{
sym_yahoo(symbols, ..., curl_options = list())

\method{import_ohlc_collection}{yahoo}(symbol_spec, dates = NULL, ...)
}
\arguments{
\item{symbols}{Ticker symbols to import.}

\item{\dots}{Arguments passed to other functions (not currently used).}

\item{curl_options}{Options passed to \pkg{curl} functions.}

\item{symbol_spec}{A \code{symbol_spec} object, with one element for each
symbol that will be imported.}

\item{dates}{An ISO-8601 string specifying the start and/or end dates.}
}
\value{
An object of class \code{symbol_spec}. The object contains a
    \code{src_attr} attribute that stores any additional details that need
    to be used for the connection to the data source.

A \code{ohlc_collection} object, with one element for each
   \code{symbol_spec}.
}
\description{
Imports data from OHLC Yahoo Finance for each symbol in \code{symbol_spec}. This
method should not be called directly. Use \code{import_ohlc_collection} with
one or more Yahoo Finance \code{symbol_spec} objects, or \code{import_ohlc}
to import one symbol.
}
\examples{

tickers <- sym_yahoo(c("AAPL", "NFLX"))

\dontrun{
# one symbol
spy <- import_ohlc(sym_yahoo("SPY"))

# multiple symbols
ohlc <- import_ohlc_collection(tickers)
}

}
\references{
\url{https://finance.yahoo.com/}
}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
