% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{import_ohlc}
\alias{import_ohlc}
\title{Import OHLC(VA) data from one or more sources}
\usage{
import_ohlc(symbol_spec, ...)
}
\arguments{
\item{symbol_spec}{A symbol specification object.}

\item{\dots}{Additional parameters passed to methods.}
}
\value{
An object of class \code{multiple_ohlc}.
}
\description{
This function imports data that has columns with open, high, low, close,
and possibly volume and/or adjusted close (if the data source returns it).
}
\examples{

tickers <- sym_yahoo(c("AAPL", "NFLX"))
ohlc <- import_ohlc(tickers)

}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
