% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ohlc_collection.R
\name{import_ohlc_collection}
\alias{import_ohlc_collection}
\title{Import a collection of OHLC data from one or more sources}
\usage{
import_ohlc_collection(symbol_spec, dates = NULL, ...)
}
\arguments{
\item{symbol_spec}{A symbol specification object.}

\item{dates}{An ISO-8601 string specifying the start and/or end dates.}

\item{\dots}{Additional parameters passed to methods.}
}
\value{
An object of class \code{ohlc_collection}.
}
\description{
This function imports data that has columns with open, high, low, close,
and possibly volume and/or adjusted close (if the data source returns it).
}
\examples{

\dontrun{
tickers <- sym_yahoo(c("AAPL", "NFLX"))
collection <- import_ohlc_collection(tickers)
}

}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
