% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generics.R
\name{import_series}
\alias{import_series}
\title{Import univariate series from one or more sources}
\usage{
import_series(symbol_spec, dates = NULL, ...)
}
\arguments{
\item{symbol_spec}{A symbol specification object.}

\item{dates}{An ISO-8601 string specifying the start and/or end dates.}

\item{\dots}{Additional parameters passed to methods.}
}
\value{
An object of class \code{xts} with one column per series.
}
\description{
This function imports data where each column represents a different series.
}
\examples{

series_symbols <- sym_fred(c("DGS10", "DGS5"))
treasury_rates <- import_series(series_symbols)

}
\author{
Joshua Ulrich
}
\keyword{IO}
\keyword{connection}
